requirements.txt
pytest>=7.0
pydantic>=2.0
PySide6>=6.10
nanpy>=0.9.6
pyserial>=3.5
opencv-python==4.13.0.92
onvif_zeep==0.2.12